change at mail templet

+++ app/Http/Controllers/DataController.php
@@ -264,17 +264,21 @@ class DataController extends Controller
     public function mail_destroy(string $id)
     {
         $email = Email::with('attachment')->find($id);
+
+        if (!$email) {
+            return redirect()->route('mail.table')->with('error', 'Email not found or already deleted.');
+        }
+
         $attachment = $email->attachment;
-        
+
         $email->delete();
-        
+
         if ($attachment) {
             Storage::disk('public')->delete($attachment->path);
             $attachment->delete();
         }
-        
-          $emails = Email::with('attachment')->paginate(25);
-      
+
+        $emails = Email::with('attachment')->paginate(25);
 
         return view('mail_table', compact('emails'));
     }
@@ -357,8 +361,8 @@ class DataController extends Controller
     public function category()
     {
         $mail_template = Email::pluck('name', 'id');
-        $categories = Category::Leftjoin('emails','emails.id','=','categories.mail_templet')
-        ->select('categories.*','emails.name as mail_templet')->get();
+        $categories = Category::Leftjoin('emails', 'emails.id', '=', 'categories.mail_templet')
+            ->select('categories.*', 'emails.name as mail_templet')->get();
         // dd($categories);
         return view('category', compact('mail_template', 'categories'));
     }
@@ -546,8 +550,7 @@ class DataController extends Controller
 
         if (Auth::guard('admin')->attempt($credentials)) {
             return redirect()->route('mail.create');
-        } 
-        else {
+        } else {
             return redirect()->back()->withErrors(['password' => 'Invalid phone number or password']);
         }
         return redirect()->route('login.view')->withErrors(['email' => 'Invalid phone number or password']);
@@ -559,7 +562,7 @@ class DataController extends Controller
         // dd($request->all());
         if (Auth::guard('admin')->check()) {
             Auth::guard('admin')->logout();
-        } 
+        }
 
 
         return redirect()->route('login.view');
